feat:ログインなしでリストURLを閲覧できるようにしました&&各ページでuseEffectでfirebaseにアクセスするように改善しました

+++ src/components/templates/MyItem.tsx
@@ -1,13 +1,13 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
 import { useSelector } from "react-redux";
 import { selectItem } from "../../features/itemSlice";
 import Button from "../atoms/Button";
 import UrlButton from "../atoms/UrlButton";
 import DeleteIcon from "@material-ui/icons/Delete";
 import { db } from "../../firebase";
-import { useHistory } from "react-router-dom";
 import { selectPost } from "../../features/postSlice";
 import { selectUser } from "../../features/userSlice";
+import { useHistory, useParams } from "react-router-dom";
 interface PROPS {
   itemName: string;
   itemPostId: string;
@@ -22,6 +22,43 @@ const MyItem: React.FC<PROPS> = (props) => {
   const history = useHistory();
   const user = useSelector(selectUser);
   const post = useSelector(selectPost);
+  const [posts, setPosts] = useState<any>("");
+  const params = useParams() as any;
+  const id = params.id as string;
+  const itemId = params.itemId as string;
+  const uid = params.uid as string;
+
+  // useEffect(() => {
+  //   const unSub = db
+  //     .collection(user.uid)
+  //     .doc(id)
+  //     .get()
+  //     .then((doc) => setPosts(doc.data()));
+
+  //   // return () => {
+  //   //   unSub();
+  //   // };
+  // }, []);
+
+  useEffect(() => {
+    const unSub = db
+      .collection(uid)
+      .doc(id)
+      // .doc(listId)
+      .collection("restaurant")
+      .doc(itemId)
+      // .doc(restaurantId)
+      .get()
+      .then((doc) => setPosts(doc.data()));
+
+    // return () => {
+    //   unSub();
+    // };
+  }, []);
+
+  // console.log(posts);
+  console.log(itemId);
+  // console.log(id);
 
   const deleteItem = () => {
     db.collection(user.uid)
@@ -35,26 +72,36 @@ const MyItem: React.FC<PROPS> = (props) => {
       .catch((error) => {
         console.error("Error removing document: ", error);
       });
-    history.push(`/lists/${post.postId}`);
+    history.push(`/${uid}/${post.postId}`);
   };
 
   return (
     <div className="flex-1">
       <p>MyItem.tsx</p>
       <p>この店を削除する</p>
-      <DeleteIcon fontSize="large" onClick={deleteItem} />
-      <p className="text-center">{item.itemName}</p>
-      <p className="text-center">{item.itemMemo}</p>
+      {user.uid && <DeleteIcon fontSize="large" onClick={deleteItem} />}
+      {/* <p className="text-center">{item.itemName}</p> */}
+      <p className="text-center">{posts.name}</p>
+      {/* <p className="text-center">{item.itemMemo}</p> */}
+      <p className="text-center">{posts.memo}</p>
       {/* <p>{item.itemPostId}</p> */}
-      <UrlButton buttonText="詳しい店情報" buttonLink={item.itemUrl} />
+      {/* <UrlButton buttonText="詳しい店情報" buttonLink={item.itemUrl} /> */}
+      <UrlButton buttonText="詳しい店情報" buttonLink={posts.url} />
       <div className="">
-        <img
+        {/* <img
           src={item.itemImageUrl}
+          className="text-center w-72 h-72 object-cover m-auto"
+          alt=""
+        /> */}
+        <img
+          src={posts.imageurl}
           className="text-center w-72 h-72 object-cover m-auto"
           alt=""
         />
       </div>
-      <p>added by {item.itemUserName}</p>
+      {/* <p>added by {item.itemUserName}</p> */}
+      {/* <p>added by {user.displayName}</p> */}
+      <p>added by {posts.username}</p>
     </div>
   );
 };
